clean code

+++ vllm/model_executor/models/transformers.py
@@ -766,7 +766,8 @@ class TransformersForMultimodalLM(TransformersForCausalLM, SupportsMultiModal):
             if isinstance(pixel_values, torch.Tensor):
                 pixel_values = flatten_bn(pixel_values).to(self.dtype)
             elif is_list_of(pixel_values, torch.Tensor):
-                pixel_values = torch.stack(flatten_bn(pixel_values)).to(self.dtype)
+                pixel_values = torch.stack(flatten_bn(pixel_values)).to(
+                    self.dtype)
             else:
                 raise ValueError(
                     f"Unsupported pixel_values type {type(pixel_values)}. "
